Fixed Django organization and prepared the app for prediction
The project is now divided into two different apps: users (which manages the authentications) and genre_classification (which manages the prediction app itself).

The prediction app is prepared so that when the prediction model will be done the predictions will be notified to the user.

Files are no longer uploaded to the project as there is no need for that, the audio files are analysed and the feature are extracted to then be feeded to the prediction model.