Implemented model in django app and changed notebook
The django app now successfully implements the keras model. When a song is uploaded a prediction is made and the genre predicted is prompted to the user.

The jupyter-notebook file now can save the encoder and the scaler as files to then use them to scale the extracted data and make a prediction and then use the label encoder to translate the predction result to labels (genres)