Created train_model.py script
The script is the adaptation of the login in the notebook for the model training.
The script takes as parameters the db and the model name, the model is trained with the data in the db and the model is saved in the /models folder.
Scaler, Label Encoder and Model are saved as files in the directory /models/{model name}