Added first data validation and fixed bug at first boot
The uploaded db for retraining is compared to the active model
to check whether they have the same features, if not the data
is not valid and the user prompted.
Fixed bug for which at the first startup the website would crash
because it could not retrieve the active model.